requirements.txt
fastapi==0.100.1
uvicorn==0.22.0
sqlalchemy==1.4.54
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pydantic>=2.0.0
python-dotenv==0.20.0
bcrypt==3.2.2
greenlet
asyncpg